pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hola-audio"
version = "0.1.0"
description = "Cross-platform voice input application with offline speech recognition"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "hery-node"},
]
keywords = ["speech-recognition", "voice-input", "asr", "nemo", "canary"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",
    "pynput>=1.7.6",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "platformdirs>=4.0.0",
    "pyperclip>=1.8.0",
]

[project.optional-dependencies]
gpu = [
    "nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git",
    "torch>=2.6.0",
]
tray = [
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]
correction = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
all = [
    "hola-audio[gpu,tray,correction,dev]",
]

[project.scripts]
hola-audio = "hola_audio.cli:main"

[project.urls]
Homepage = "https://github.com/hery-node/hola-audio"
Repository = "https://github.com/hery-node/hola-audio"
Issues = "https://github.com/hery-node/hola-audio/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
